Updates mark scheme for 24/25 for viva and background
Two main updates for the mark scheme for 24/25, these concern adding in
more specific aspects of the viva and the contextual background section
within the dissertation.

* **Contextual background**: this change aims to solidify that students need
  to have a literature review within their project within the report
presentation pillar (pillar 4) _in addition_ to having cited prior work
(which could have been in a method section or evaluation section). This
pillar was also re-ordered so that in each grade boundary lit. review is
mentioned first and then prior work.
* **viva**: this change adds a specific marking criteria of the viva to
  do with the student's performance in the viva beyond answering
questions to their understanding of the project. A student that doesn't
understand large aspects of their project should not pass the unit and
this criteria is updated as such.